+++ src/Services/GoogleAds/ReportDownload.php
@@ -64,6 +64,26 @@ class ReportDownload
         $header = explode(',',$csv[0]);
         $csv    = array_filter($csv);
 
+        $header = [];
+        foreach(explode(',',$csv[0]) as $label)
+        {
+            $label = strtolower($label);
+
+            switch($label) {
+                case 'day' : $label = 'date'; break;
+                case 'campaign id' : $label = 'campaign_id'; break;
+                case 'campaign state' : $label = 'campaign_status'; break;
+                case 'campaign' : $label = 'campaign_name'; break;
+                case 'ad group' : $label = 'ad_group_name'; break;
+                case 'advertising channel' : $label = 'channel'; break;
+                case 'ad group id' : $label = 'ad_group_id'; break;
+                case 'total conv. value' : $label = 'conversion_value'; break;
+                default :
+            }
+
+            $header[] = str_replace(' ','_',$label);
+        }
+
         unset($csv[0]);
 
         $report = [];
@@ -78,7 +98,7 @@ class ReportDownload
 
                 $n = $header[$index2];
 
-                if ($n == 'Cost') $cs = round( intval($cs) / 1000000,2);
+                if ($n == 'cost') $cs = round( intval($cs) / 1000000,2);
 
                 $r[$n] = $cs;
             }