+++ src/Services/GoogleAds/Reports.php
@@ -138,8 +138,8 @@ class Reports
             'Cost',
             'Conversions',
             'ConversionValue',
-            'AllConversions',
-            'AllConversionValue'
+            // 'AllConversions',
+            // 'AllConversionValue'
         ],true);
 
         $reportDefinition = $this->reportDefinition(ReportDefinitionReportType::ACCOUNT_PERFORMANCE_REPORT);
@@ -167,8 +167,8 @@ class Reports
             'Cost',
             'Conversions',
             'ConversionValue',
-            'AllConversions',
-            'AllConversionValue'
+            // 'AllConversions',
+            // 'AllConversionValue'
         ],true);
 
         $reportDefinition = $this->reportDefinition(ReportDefinitionReportType::CAMPAIGN_PERFORMANCE_REPORT);
@@ -195,8 +195,8 @@ class Reports
             'Cost',
             'Conversions',
             'ConversionValue',
-            'AllConversions',
-            'AllConversionValue'
+            // 'AllConversions',
+            // 'AllConversionValue'
         ],true);
 
         $reportDefinition = $this->reportDefinition(ReportDefinitionReportType::ADGROUP_PERFORMANCE_REPORT);